storage: separate files for read/write path
Release note: None
