+++ pkg/storage/replica.go
@@ -18,7 +18,6 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"math"
 	"math/rand"
 	"os"
 	"reflect"
@@ -36,8 +35,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/storage/abortspan"
 	"github.com/cockroachdb/cockroach/pkg/storage/batcheval"
 	"github.com/cockroachdb/cockroach/pkg/storage/batcheval/result"
-	"github.com/cockroachdb/cockroach/pkg/storage/closedts/ctpb"
-	ctstorage "github.com/cockroachdb/cockroach/pkg/storage/closedts/storage"
 	"github.com/cockroachdb/cockroach/pkg/storage/engine"
 	"github.com/cockroachdb/cockroach/pkg/storage/engine/enginepb"
 	"github.com/cockroachdb/cockroach/pkg/storage/rangefeed"
@@ -48,7 +45,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/storage/storagebase"
 	"github.com/cockroachdb/cockroach/pkg/storage/storagepb"
 	"github.com/cockroachdb/cockroach/pkg/storage/txnwait"
-	"github.com/cockroachdb/cockroach/pkg/util"
 	"github.com/cockroachdb/cockroach/pkg/util/envutil"
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/humanizeutil"
@@ -2285,361 +2281,6 @@ func (r *Replica) maybeWatchForMerge(ctx context.Context) error {
 	return err
 }
 
-// executeReadOnlyBatch updates the read timestamp cache and waits for any
-// overlapping writes currently processing through Raft ahead of us to
-// clear via the latches.
-func (r *Replica) executeReadOnlyBatch(
-	ctx context.Context, ba roachpb.BatchRequest,
-) (br *roachpb.BatchResponse, pErr *roachpb.Error) {
-	// If the read is not inconsistent, the read requires the range lease or
-	// permission to serve via follower reads.
-	var status storagepb.LeaseStatus
-	if ba.ReadConsistency.RequiresReadLease() {
-		if status, pErr = r.redirectOnOrAcquireLease(ctx); pErr != nil {
-			if lErr, ok := pErr.GetDetail().(*roachpb.NotLeaseHolderError); ok &&
-				FollowerReadsEnabled.Get(&r.store.cfg.Settings.SV) &&
-				lErr.LeaseHolder != nil && lErr.Lease.Type() == roachpb.LeaseEpoch {
-
-				r.mu.RLock()
-				lai := r.mu.state.LeaseAppliedIndex
-				r.mu.RUnlock()
-				if !r.store.cfg.ClosedTimestamp.Provider.CanServe(
-					lErr.LeaseHolder.NodeID, ba.Timestamp, r.RangeID, ctpb.Epoch(lErr.Lease.Epoch), ctpb.LAI(lai),
-				) {
-					r.store.cfg.ClosedTimestamp.Clients.Request(lErr.LeaseHolder.NodeID, r.RangeID)
-					// NB: this can't go behind V(x) because the log message created by the
-					// storage might be gigantic in real clusters, and we don't want to trip it
-					// using logspy.
-					if false {
-						log.Warningf(ctx, "can't serve follower read for %s at epo %d lai %d, storage is %s",
-							ba.Timestamp, lErr.Lease.Epoch, lai,
-							r.store.cfg.ClosedTimestamp.Storage.(*ctstorage.MultiStorage).StringForNodes(lErr.LeaseHolder.NodeID),
-						)
-					}
-					return nil, pErr
-				}
-				// This replica can serve this read!
-				//
-				// TODO(tschottdorf): once a read for a timestamp T has been served, the replica may
-				// serve reads for that and smaller timestamps forever.
-				log.Event(ctx, "serving via follower read")
-			} else {
-				return nil, pErr
-			}
-		}
-	}
-	r.limitTxnMaxTimestamp(ctx, &ba, status)
-
-	spans, err := r.collectSpans(&ba)
-	if err != nil {
-		return nil, roachpb.NewError(err)
-	}
-
-	// Acquire latches to prevent overlapping commands from executing
-	// until this command completes.
-	log.Event(ctx, "acquire latches")
-	endCmds, err := r.beginCmds(ctx, &ba, spans)
-	if err != nil {
-		return nil, roachpb.NewError(err)
-	}
-
-	log.Event(ctx, "waiting for read lock")
-	r.readOnlyCmdMu.RLock()
-	defer r.readOnlyCmdMu.RUnlock()
-
-	// Guarantee we release the latches that we just acquired. It is
-	// important that this is inside the readOnlyCmdMu lock so that the
-	// timestamp cache update is synchronized. This is wrapped to delay
-	// pErr evaluation to its value when returning.
-	defer func() {
-		endCmds.done(br, pErr, proposalNoReevaluation)
-	}()
-
-	// TODO(nvanbenschoten): Can this be moved into Replica.requestCanProceed?
-	if _, err := r.IsDestroyed(); err != nil {
-		return nil, roachpb.NewError(err)
-	}
-
-	rSpan, err := keys.Range(ba)
-	if err != nil {
-		return nil, roachpb.NewError(err)
-	}
-
-	if err := r.requestCanProceed(rSpan, ba.Timestamp); err != nil {
-		return nil, roachpb.NewError(err)
-	}
-
-	// Evaluate read-only batch command. It checks for matching key range; note
-	// that holding readOnlyCmdMu throughout is important to avoid reads from the
-	// "wrong" key range being served after the range has been split.
-	var result result.Result
-	rec := NewReplicaEvalContext(r, spans)
-	readOnly := r.store.Engine().NewReadOnly()
-	if util.RaceEnabled {
-		readOnly = spanset.NewReadWriter(readOnly, spans)
-	}
-	defer readOnly.Close()
-	br, result, pErr = evaluateBatch(ctx, storagebase.CmdIDKey(""), readOnly, rec, nil, ba)
-
-	if result.Local.DetachMaybeWatchForMerge() {
-		if err := r.maybeWatchForMerge(ctx); err != nil {
-			return nil, roachpb.NewError(err)
-		}
-	}
-
-	if intents := result.Local.DetachIntents(); len(intents) > 0 {
-		log.Eventf(ctx, "submitting %d intents to asynchronous processing", len(intents))
-		// We only allow synchronous intent resolution for consistent requests.
-		// Intent resolution is async/best-effort for inconsistent requests.
-		//
-		// An important case where this logic is necessary is for RangeLookup
-		// requests. In their case, synchronous intent resolution can deadlock
-		// if the request originated from the local node which means the local
-		// range descriptor cache has an in-flight RangeLookup request which
-		// prohibits any concurrent requests for the same range. See #17760.
-		allowSyncProcessing := ba.ReadConsistency == roachpb.CONSISTENT
-		if err := r.store.intentResolver.cleanupIntentsAsync(ctx, r, intents, allowSyncProcessing); err != nil {
-			log.Warning(ctx, err)
-		}
-	}
-	if pErr != nil {
-		log.VErrEvent(ctx, 3, pErr.String())
-	} else {
-		log.Event(ctx, "read completed")
-	}
-	return br, pErr
-}
-
-// executeWriteBatch is the entry point for client requests which may mutate the
-// range's replicated state. Requests taking this path are evaluated and ultimately
-// serialized through Raft, but pass through additional machinery whose goal is
-// to allow commands which commute to be proposed in parallel. The naive
-// alternative, submitting requests to Raft one after another, paying massive
-// latency, is only taken for commands whose effects may overlap.
-//
-// Internally, multiple iterations of the above process may take place
-// due to the Raft proposal failing retryably, possibly due to proposal
-// reordering or re-proposals. We call these retry "re-evaluations" since the
-// request is evaluated again (against a fresh engine snapshot).
-func (r *Replica) executeWriteBatch(
-	ctx context.Context, ba roachpb.BatchRequest,
-) (*roachpb.BatchResponse, *roachpb.Error) {
-	for {
-		// TODO(andrei): export some metric about re-evaluations.
-		br, pErr, retry := r.tryExecuteWriteBatch(ctx, ba)
-		if retry == proposalIllegalLeaseIndex {
-			log.VEventf(ctx, 2, "retry: proposalIllegalLeaseIndex")
-			if pErr != nil {
-				log.Fatalf(ctx, "both error and retry returned: %s", pErr)
-			}
-			continue // retry
-		}
-		return br, pErr
-	}
-}
-
-// tryExecuteWriteBatch is invoked by executeWriteBatch, which will
-// call this method until it returns a non-retryable result (i.e. no
-// proposalRetryReason is returned).
-//
-// Concretely,
-//
-// - Latches for the keys affected by the command are acquired (i.e.
-//   tracked as in-flight mutations).
-// - In doing so, we wait until no overlapping mutations are in flight.
-// - The timestamp cache is checked to determine if the command's affected keys
-//   were accessed with a timestamp exceeding that of the command; if so, the
-//   command's timestamp is incremented accordingly.
-// - A RaftCommand is constructed. If proposer-evaluated KV is active,
-//   the request is evaluated and the Result is placed in the
-//   RaftCommand. If not, the request itself is added to the command.
-// - The proposal is inserted into the Replica's in-flight proposals map,
-//   a lease index is assigned to it, and it is submitted to Raft, returning
-//   a channel.
-// - The result of the Raft proposal is read from the channel and the command
-//   registered with the timestamp cache, its latches are released, and
-//   its result (which could be an error) is returned to the client.
-//
-// Returns exactly one of a response, an error or re-evaluation reason.
-//
-// NB: changing BatchRequest to a pointer here would have to be done cautiously
-// as this method makes the assumption that it operates on a shallow copy (see
-// call to applyTimestampCache).
-func (r *Replica) tryExecuteWriteBatch(
-	ctx context.Context, ba roachpb.BatchRequest,
-) (br *roachpb.BatchResponse, pErr *roachpb.Error, retry proposalReevaluationReason) {
-	startTime := timeutil.Now()
-
-	if err := r.maybeBackpressureWriteBatch(ctx, ba); err != nil {
-		return nil, roachpb.NewError(err), proposalNoReevaluation
-	}
-
-	spans, err := r.collectSpans(&ba)
-	if err != nil {
-		return nil, roachpb.NewError(err), proposalNoReevaluation
-	}
-
-	var endCmds *endCmds
-	if !ba.IsLeaseRequest() {
-		// Acquire latches to prevent overlapping commands from executing until
-		// this command completes. Note that this must be done before getting
-		// the max timestamp for the key(s), as timestamp cache is only updated
-		// after preceding commands have been run to successful completion.
-		log.Event(ctx, "acquire latches")
-		var err error
-		endCmds, err = r.beginCmds(ctx, &ba, spans)
-		if err != nil {
-			return nil, roachpb.NewError(err), proposalNoReevaluation
-		}
-	}
-
-	// Guarantee we release the latches that we just acquired. This is
-	// wrapped to delay pErr evaluation to its value when returning.
-	defer func() {
-		if endCmds != nil {
-			endCmds.done(br, pErr, retry)
-		}
-	}()
-
-	var lease roachpb.Lease
-	var status storagepb.LeaseStatus
-	// For lease commands, use the provided previous lease for verification.
-	if ba.IsSingleSkipLeaseCheckRequest() {
-		lease = ba.GetPrevLeaseForLeaseRequest()
-	} else {
-		// Other write commands require that this replica has the range
-		// lease.
-		if status, pErr = r.redirectOnOrAcquireLease(ctx); pErr != nil {
-			return nil, pErr, proposalNoReevaluation
-		}
-		lease = status.Lease
-	}
-	r.limitTxnMaxTimestamp(ctx, &ba, status)
-
-	minTS, untrack := r.store.cfg.ClosedTimestamp.Tracker.Track(ctx)
-	defer untrack(ctx, 0, 0) // covers all error returns below
-
-	// Examine the read and write timestamp caches for preceding
-	// commands which require this command to move its timestamp
-	// forward. Or, in the case of a transactional write, the txn
-	// timestamp and possible write-too-old bool.
-	if bumped, pErr := r.applyTimestampCache(ctx, &ba, minTS); pErr != nil {
-		return nil, pErr, proposalNoReevaluation
-	} else if bumped {
-		// If we bump the transaction's timestamp, we must absolutely
-		// tell the client in a response transaction (for otherwise it
-		// doesn't know about the incremented timestamp). Response
-		// transactions are set far away from this code, but at the time
-		// of writing, they always seem to be set. Since that is a
-		// likely target of future micro-optimization, this assertion is
-		// meant to protect against future correctness anomalies.
-		defer func() {
-			if br != nil && ba.Txn != nil && br.Txn == nil {
-				log.Fatalf(ctx, "assertion failed: transaction updated by "+
-					"timestamp cache, but transaction returned in response; "+
-					"updated timestamp would have been lost (recovered): "+
-					"%s in batch %s", ba.Txn, ba,
-				)
-			}
-		}()
-	}
-
-	log.Event(ctx, "applied timestamp cache")
-
-	ch, tryAbandon, maxLeaseIndex, pErr := r.propose(ctx, lease, ba, endCmds, spans)
-	if pErr != nil {
-		if maxLeaseIndex != 0 {
-			log.Fatalf(
-				ctx, "unexpected max lease index %d assigned to failed proposal: %s, error %s",
-				maxLeaseIndex, ba, pErr,
-			)
-		}
-		return nil, pErr, proposalNoReevaluation
-	}
-	// A max lease index of zero is returned when no proposal was made or a lease was proposed.
-	// In both cases, we don't need to communicate a MLAI.
-	if maxLeaseIndex != 0 {
-		untrack(ctx, r.RangeID, ctpb.LAI(maxLeaseIndex))
-	}
-
-	// After the command is proposed to Raft, invoking endCmds.done is now the
-	// responsibility of processRaftCommand.
-	endCmds = nil
-
-	// If the command was accepted by raft, wait for the range to apply it.
-	ctxDone := ctx.Done()
-	shouldQuiesce := r.store.stopper.ShouldQuiesce()
-	slowTimer := timeutil.NewTimer()
-	defer slowTimer.Stop()
-	slowTimer.Reset(base.SlowRequestThreshold)
-	tBegin := timeutil.Now()
-
-	for {
-		select {
-		case propResult := <-ch:
-			// Semi-synchronously process any intents that need resolving here in
-			// order to apply back pressure on the client which generated them. The
-			// resolution is semi-synchronous in that there is a limited number of
-			// outstanding asynchronous resolution tasks allowed after which
-			// further calls will block.
-			if len(propResult.Intents) > 0 {
-				// TODO(peter): Re-proposed and canceled (but executed) commands can
-				// both leave intents to GC that don't hit this code path. No good
-				// solution presents itself at the moment and such intents will be
-				// resolved on reads.
-				if err := r.store.intentResolver.cleanupIntentsAsync(ctx, r, propResult.Intents, true /* allowSync */); err != nil {
-					log.Warning(ctx, err)
-				}
-			}
-			if len(propResult.EndTxns) > 0 {
-				if err := r.store.intentResolver.cleanupTxnIntentsAsync(ctx, r, propResult.EndTxns, true /* allowSync */); err != nil {
-					log.Warning(ctx, err)
-				}
-			}
-			return propResult.Reply, propResult.Err, propResult.ProposalRetry
-		case <-slowTimer.C:
-			slowTimer.Read = true
-			log.Warningf(ctx, "have been waiting %s for proposing command %s",
-				base.SlowRequestThreshold, ba)
-			r.store.metrics.SlowRaftRequests.Inc(1)
-			defer func() {
-				r.store.metrics.SlowRaftRequests.Dec(1)
-				contextStr := ""
-				if err := ctx.Err(); err != nil {
-					contextStr = " with context cancellation"
-				}
-				log.Infof(ctx, "slow command %s finished after %s%s", ba, timeutil.Since(tBegin), contextStr)
-			}()
-
-		case <-ctxDone:
-			// If our context was canceled, return an AmbiguousResultError
-			// if the command isn't already being executed and using our
-			// context, in which case we expect it to finish soon. The
-			// AmbiguousResultError indicates to caller that the command may
-			// have executed.
-			if tryAbandon() {
-				log.VEventf(ctx, 2, "context cancellation after %0.1fs of attempting command %s",
-					timeutil.Since(startTime).Seconds(), ba)
-				return nil, roachpb.NewError(roachpb.NewAmbiguousResultError(ctx.Err().Error())), proposalNoReevaluation
-			}
-			ctxDone = nil
-		case <-shouldQuiesce:
-			// If shutting down, return an AmbiguousResultError if the
-			// command isn't already being executed and using our context,
-			// in which case we expect it to finish soon. AmbiguousResultError
-			// indicates to caller that the command may have executed. If
-			// tryAbandon fails, we iterate through the loop again to wait
-			// for the command to finish.
-			if tryAbandon() {
-				log.VEventf(ctx, 2, "shutdown cancellation after %0.1fs of attempting command %s",
-					timeutil.Since(startTime).Seconds(), ba)
-				return nil, roachpb.NewError(roachpb.NewAmbiguousResultError("server shutdown")), proposalNoReevaluation
-			}
-			shouldQuiesce = nil
-		}
-	}
-}
-
 // requestToProposal converts a BatchRequest into a ProposalData, by
 // evaluating it. The returned ProposalData is partially valid even
 // on a non-nil *roachpb.Error and should be proposed through Raft
@@ -3203,526 +2844,6 @@ func checkIfTxnAborted(
 			roachpb.NewTransactionAbortedError(roachpb.ABORT_REASON_ABORT_SPAN), &newTxn)
 	}
 	return nil
-}
-
-// evaluateWriteBatch evaluates the supplied batch.
-//
-// If the batch is transactional and has all the hallmarks of a 1PC
-// commit (i.e. includes BeginTransaction & EndTransaction, and
-// there's nothing to suggest that the transaction will require retry
-// or restart), the batch's txn is stripped and it's executed as an
-// atomic batch write. If the writes cannot all be completed at the
-// intended timestamp, the batch's txn is restored and it's
-// re-executed in full. This allows it to lay down intents and return
-// an appropriate retryable error.
-func (r *Replica) evaluateWriteBatch(
-	ctx context.Context, idKey storagebase.CmdIDKey, ba roachpb.BatchRequest, spans *spanset.SpanSet,
-) (engine.Batch, enginepb.MVCCStats, *roachpb.BatchResponse, result.Result, *roachpb.Error) {
-	ms := enginepb.MVCCStats{}
-	// If not transactional or there are indications that the batch's txn will
-	// require restart or retry, execute as normal.
-	if isOnePhaseCommit(ba, r.store.TestingKnobs()) {
-		_, hasBegin := ba.GetArg(roachpb.BeginTransaction)
-		arg, _ := ba.GetArg(roachpb.EndTransaction)
-		etArg := arg.(*roachpb.EndTransactionRequest)
-
-		// Try executing with transaction stripped. We use the transaction timestamp
-		// to write any values as it may have been advanced by the timestamp cache.
-		strippedBa := ba
-		strippedBa.Timestamp = strippedBa.Txn.Timestamp
-		strippedBa.Txn = nil
-		if hasBegin {
-			strippedBa.Requests = ba.Requests[1 : len(ba.Requests)-1] // strip begin/end txn reqs
-		} else {
-			strippedBa.Requests = ba.Requests[:len(ba.Requests)-1] // strip end txn req
-		}
-
-		// If there were no refreshable spans earlier in the txn
-		// (e.g. earlier gets or scans), then the batch can be retried
-		// locally in the event of write too old errors.
-		retryLocally := etArg.NoRefreshSpans && !ba.Txn.OrigTimestampWasObserved
-
-		// If all writes occurred at the intended timestamp, we've succeeded on the fast path.
-		rec := NewReplicaEvalContext(r, spans)
-		batch, br, res, pErr := r.evaluateWriteBatchWithLocalRetries(
-			ctx, idKey, rec, &ms, strippedBa, spans, retryLocally,
-		)
-		if pErr == nil && (ba.Timestamp == br.Timestamp ||
-			(retryLocally && !batcheval.IsEndTransactionExceedingDeadline(br.Timestamp, *etArg))) {
-			clonedTxn := ba.Txn.Clone()
-			clonedTxn.Status = roachpb.COMMITTED
-			// Make sure the returned txn has the actual commit
-			// timestamp. This can be different if the stripped batch was
-			// executed at the server's hlc now timestamp.
-			clonedTxn.Timestamp = br.Timestamp
-
-			// If the end transaction is not committed, clear the batch and mark the status aborted.
-			if !etArg.Commit {
-				clonedTxn.Status = roachpb.ABORTED
-				batch.Close()
-				batch = r.store.Engine().NewBatch()
-				ms = enginepb.MVCCStats{}
-			} else {
-				// Run commit trigger manually.
-				innerResult, err := batcheval.RunCommitTrigger(ctx, rec, batch, &ms, *etArg, &clonedTxn)
-				if err != nil {
-					return batch, ms, br, res, roachpb.NewErrorf("failed to run commit trigger: %s", err)
-				}
-				if err := res.MergeAndDestroy(innerResult); err != nil {
-					return batch, ms, br, res, roachpb.NewError(err)
-				}
-			}
-
-			br.Txn = &clonedTxn
-			// Add placeholder responses for begin & end transaction requests.
-			var resps []roachpb.ResponseUnion
-			if hasBegin {
-				resps = make([]roachpb.ResponseUnion, len(br.Responses)+2)
-				resps[0].MustSetInner(&roachpb.BeginTransactionResponse{})
-				copy(resps[1:], br.Responses)
-			} else {
-				resps = append(br.Responses, roachpb.ResponseUnion{})
-			}
-			resps[len(resps)-1].MustSetInner(&roachpb.EndTransactionResponse{OnePhaseCommit: true})
-			br.Responses = resps
-			return batch, ms, br, res, nil
-		}
-
-		ms = enginepb.MVCCStats{}
-
-		// Handle the case of a required one phase commit transaction.
-		if etArg.Require1PC {
-			if pErr != nil {
-				return batch, ms, nil, result.Result{}, pErr
-			} else if ba.Timestamp != br.Timestamp {
-				err := roachpb.NewTransactionRetryError(roachpb.RETRY_REASON_UNKNOWN)
-				return batch, ms, nil, result.Result{}, roachpb.NewError(err)
-			}
-			log.Fatal(ctx, "unreachable")
-		}
-
-		batch.Close()
-		log.VEventf(ctx, 2, "1PC execution failed, reverting to regular execution for batch")
-	}
-
-	rec := NewReplicaEvalContext(r, spans)
-	// We can retry locally if this is a non-transactional request.
-	canRetry := ba.Txn == nil
-	batch, br, res, pErr := r.evaluateWriteBatchWithLocalRetries(ctx, idKey, rec, &ms, ba, spans, canRetry)
-	return batch, ms, br, res, pErr
-}
-
-// evaluateWriteBatchWithLocalRetries invokes evaluateBatch and
-// retries in the event of a WriteTooOldError at a higher timestamp if
-// canRetry is true.
-func (r *Replica) evaluateWriteBatchWithLocalRetries(
-	ctx context.Context,
-	idKey storagebase.CmdIDKey,
-	rec batcheval.EvalContext,
-	ms *enginepb.MVCCStats,
-	ba roachpb.BatchRequest,
-	spans *spanset.SpanSet,
-	canRetry bool,
-) (batch engine.Batch, br *roachpb.BatchResponse, res result.Result, pErr *roachpb.Error) {
-	for retries := 0; ; retries++ {
-		if batch != nil {
-			batch.Close()
-		}
-		batch = r.store.Engine().NewBatch()
-		var opLogger *engine.OpLoggerBatch
-		if RangefeedEnabled.Get(&r.store.cfg.Settings.SV) {
-			// TODO(nvanbenschoten): once we get rid of the RangefeedEnabled
-			// cluster setting we'll need a way to turn this on when any
-			// replica (not just the leaseholder) wants it and off when no
-			// replicas want it. This turns out to be pretty involved.
-			//
-			// The current plan is to:
-			// - create a range-id local key that stores all replicas that are
-			//   subscribed to logical operations, along with their corresponding
-			//   liveness epoch.
-			// - create a new command that adds or subtracts replicas from this
-			//   structure. The command will be a write across the entire replica
-			//   span so that it is serialized with all writes.
-			// - each replica will add itself to this set when it first needs
-			//   logical ops. It will then wait until it sees the replicated command
-			//   that added itself pop out through Raft so that it knows all
-			//   commands that are missing logical ops are gone.
-			// - It will then proceed as normal, relying on the logical ops to
-			//   always be included on the raft commands. When its no longer
-			//   needs logical ops, it will remove itself from the set.
-			// - The leaseholder will have a new queue to detect registered
-			//   replicas that are no longer live and remove them from the
-			//   set to prevent "leaking" subscriptions.
-			// - The condition here to add logical logging will be:
-			//     if len(replicaState.logicalOpsSubs) > 0 { ... }
-			//
-			// An alternative to this is the reduce the cost of the including
-			// the logical op log to a negligible amount such that it can be
-			// included on all raft commands, regardless of whether any replica
-			// has a rangefeed running or not.
-			//
-			// Another alternative is to make the setting table/zone-scoped
-			// instead of a fine-grained per-replica state.
-			opLogger = engine.NewOpLoggerBatch(batch)
-			batch = opLogger
-		}
-		if util.RaceEnabled {
-			batch = spanset.NewBatch(batch, spans)
-		}
-
-		br, res, pErr = evaluateBatch(ctx, idKey, batch, rec, ms, ba)
-		// If we can retry, set a higher batch timestamp and continue.
-		if wtoErr, ok := pErr.GetDetail().(*roachpb.WriteTooOldError); ok && canRetry {
-			// Allow one retry only; a non-txn batch containing overlapping
-			// spans will always experience WriteTooOldError.
-			if retries == 1 {
-				break
-			}
-			ba.Timestamp = wtoErr.ActualTimestamp
-			continue
-		}
-		if opLogger != nil {
-			res.LogicalOpLog = &storagepb.LogicalOpLog{
-				Ops: opLogger.LogicalOps(),
-			}
-		}
-		break
-	}
-	return
-}
-
-// isOnePhaseCommit returns true iff the BatchRequest contains all commands in
-// the transaction, starting with BeginTransaction and ending with
-// EndTransaction. One phase commits are disallowed if (1) the transaction has
-// already been flagged with a write too old error, or (2) if isolation is
-// serializable and the commit timestamp has been forwarded, or (3) the
-// transaction exceeded its deadline, or (4) the testing knobs disallow optional
-// one phase commits and the BatchRequest does not require one phase commit.
-func isOnePhaseCommit(ba roachpb.BatchRequest, knobs *StoreTestingKnobs) bool {
-	if ba.Txn == nil {
-		return false
-	}
-	if !ba.IsCompleteTransaction() {
-		return false
-	}
-	arg, _ := ba.GetArg(roachpb.EndTransaction)
-	etArg := arg.(*roachpb.EndTransactionRequest)
-	if batcheval.IsEndTransactionExceedingDeadline(ba.Txn.Timestamp, *etArg) {
-		return false
-	}
-	if retry, _ := batcheval.IsEndTransactionTriggeringRetryError(ba.Txn, *etArg); retry {
-		return false
-	}
-	return !knobs.DisableOptional1PC || etArg.Require1PC
-}
-
-// optimizePuts searches for contiguous runs of Put & CPut commands in
-// the supplied request union. Any run which exceeds a minimum length
-// threshold employs a full order iterator to determine whether the
-// range of keys being written is empty. If so, then the run can be
-// set to put "blindly", meaning no iterator need be used to read
-// existing values during the MVCC write.
-// The caller should use the returned slice (which is either equal to
-// the input slice, or has been shallow-copied appropriately to avoid
-// mutating the original requests).
-func optimizePuts(
-	batch engine.ReadWriter, origReqs []roachpb.RequestUnion, distinctSpans bool,
-) []roachpb.RequestUnion {
-	var minKey, maxKey roachpb.Key
-	var unique map[string]struct{}
-	if !distinctSpans {
-		unique = make(map[string]struct{}, len(origReqs))
-	}
-	// Returns false on occurrence of a duplicate key.
-	maybeAddPut := func(key roachpb.Key) bool {
-		// Note that casting the byte slice key to a string does not allocate.
-		if unique != nil {
-			if _, ok := unique[string(key)]; ok {
-				return false
-			}
-			unique[string(key)] = struct{}{}
-		}
-		if minKey == nil || bytes.Compare(key, minKey) < 0 {
-			minKey = key
-		}
-		if maxKey == nil || bytes.Compare(key, maxKey) > 0 {
-			maxKey = key
-		}
-		return true
-	}
-
-	firstUnoptimizedIndex := len(origReqs)
-	for i, r := range origReqs {
-		switch t := r.GetInner().(type) {
-		case *roachpb.PutRequest:
-			if maybeAddPut(t.Key) {
-				continue
-			}
-		case *roachpb.ConditionalPutRequest:
-			if maybeAddPut(t.Key) {
-				continue
-			}
-		case *roachpb.InitPutRequest:
-			if maybeAddPut(t.Key) {
-				continue
-			}
-		}
-		firstUnoptimizedIndex = i
-		break
-	}
-
-	if firstUnoptimizedIndex < optimizePutThreshold { // don't bother if below this threshold
-		return origReqs
-	}
-	iter := batch.NewIterator(engine.IterOptions{
-		// We want to include maxKey in our scan. Since UpperBound is exclusive, we
-		// need to set it to the key after maxKey.
-		UpperBound: maxKey.Next(),
-	})
-	defer iter.Close()
-
-	// If there are enough puts in the run to justify calling seek,
-	// we can determine whether any part of the range being written
-	// is "virgin" and set the puts to write blindly.
-	// Find the first non-empty key in the run.
-	iter.Seek(engine.MakeMVCCMetadataKey(minKey))
-	var iterKey roachpb.Key
-	if ok, err := iter.Valid(); err != nil {
-		// TODO(bdarnell): return an error here instead of silently
-		// running without the optimization?
-		log.Errorf(context.TODO(), "Seek returned error; disabling blind-put optimization: %s", err)
-		return origReqs
-	} else if ok && bytes.Compare(iter.Key().Key, maxKey) <= 0 {
-		iterKey = iter.Key().Key
-	}
-	// Set the prefix of the run which is being written to virgin
-	// keyspace to "blindly" put values.
-	reqs := append([]roachpb.RequestUnion(nil), origReqs...)
-	for i := range reqs[:firstUnoptimizedIndex] {
-		inner := reqs[i].GetInner()
-		if iterKey == nil || bytes.Compare(iterKey, inner.Header().Key) > 0 {
-			switch t := inner.(type) {
-			case *roachpb.PutRequest:
-				shallow := *t
-				shallow.Blind = true
-				reqs[i].MustSetInner(&shallow)
-			case *roachpb.ConditionalPutRequest:
-				shallow := *t
-				shallow.Blind = true
-				reqs[i].MustSetInner(&shallow)
-			case *roachpb.InitPutRequest:
-				shallow := *t
-				shallow.Blind = true
-				reqs[i].MustSetInner(&shallow)
-			default:
-				log.Fatalf(context.TODO(), "unexpected non-put request: %s", t)
-			}
-		}
-	}
-	return reqs
-}
-
-// evaluateBatch evaluates a batch request by splitting it up into its
-// individual commands, passing them to evaluateCommand, and combining
-// the results.
-func evaluateBatch(
-	ctx context.Context,
-	idKey storagebase.CmdIDKey,
-	batch engine.ReadWriter,
-	rec batcheval.EvalContext,
-	ms *enginepb.MVCCStats,
-	ba roachpb.BatchRequest,
-) (*roachpb.BatchResponse, result.Result, *roachpb.Error) {
-	br := ba.CreateReply()
-
-	maxKeys := int64(math.MaxInt64)
-	if ba.Header.MaxSpanRequestKeys != 0 {
-		// We have a batch of requests with a limit. We keep track of how many
-		// remaining keys we can touch.
-		maxKeys = ba.Header.MaxSpanRequestKeys
-	}
-
-	// Optimize any contiguous sequences of put and conditional put ops.
-	if len(ba.Requests) >= optimizePutThreshold {
-		ba.Requests = optimizePuts(batch, ba.Requests, ba.Header.DistinctSpans)
-	}
-
-	// Create a shallow clone of the transaction. We only modify a few
-	// non-pointer fields (Sequence, DeprecatedBatchIndex, WriteTooOld, Timestamp),
-	// so this saves a few allocs.
-	if ba.Txn != nil {
-		txnShallow := *ba.Txn
-		ba.Txn = &txnShallow
-
-		// Check whether this transaction has been aborted, if applicable.
-		// This applies to writes that leave intents (the use of the
-		// IsTransactionWrite flag excludes operations like HeartbeatTxn),
-		// and reads that occur in a transaction that has already written
-		// (see #2231 for more about why we check for aborted transactions
-		// on reads). Note that 1PC transactions have had their
-		// transaction field cleared by this point so we do not execute
-		// this check in that case.
-		if ba.IsTransactionWrite() || ba.Txn.Writing {
-			// We don't check the abort span for a couple of special requests:
-			// - if the request is asking to abort the transaction, then don't check the
-			// AbortSpan; we don't want the request to be rejected if the transaction
-			// has already been aborted.
-			// - heartbeats don't check the abort span. If the txn is aborted, they'll
-			// return an aborted proto in their otherwise successful response.
-			singleAbort := ba.IsSingleEndTransactionRequest() &&
-				!ba.Requests[0].GetInner().(*roachpb.EndTransactionRequest).Commit
-			if !singleAbort && !ba.IsSingleHeartbeatTxnRequest() {
-				if pErr := checkIfTxnAborted(ctx, rec, batch, *ba.Txn); pErr != nil {
-					return nil, result.Result{}, pErr
-				}
-			}
-		}
-	}
-
-	var result result.Result
-	var writeTooOldErr *roachpb.Error
-	returnWriteTooOldErr := true
-
-	for index, union := range ba.Requests {
-		// Execute the command.
-		args := union.GetInner()
-		if ba.Txn != nil {
-			// Sequence numbers used to be set on each BatchRequest instead of
-			// on each individual Request. This meant that all Requests in a
-			// BatchRequest shared the same sequence number, so a BatchIndex was
-			// augmented to provide an ordering between them. Individual
-			// Requests were later given their own sequence numbers, so the
-			// BatchIndex was no longer necessary.
-			if seqNum := args.Header().Sequence; seqNum != 0 {
-				// Set the Request's sequence number on the TxnMeta for this
-				// request. Each request will set their own sequence number on
-				// the TxnMeta, which is stored as part of an intent.
-				ba.Txn.Sequence = seqNum
-			}
-		}
-		// Note that responses are populated even when an error is returned.
-		// TODO(tschottdorf): Change that. IIRC there is nontrivial use of it currently.
-		reply := br.Responses[index].GetInner()
-		curResult, pErr := evaluateCommand(ctx, idKey, index, batch, rec, ms, ba.Header, maxKeys, args, reply)
-
-		if err := result.MergeAndDestroy(curResult); err != nil {
-			// TODO(tschottdorf): see whether we really need to pass nontrivial
-			// Result up on error and if so, formalize that.
-			log.Fatalf(
-				ctx,
-				"unable to absorb Result: %s\ndiff(new, old): %s",
-				err, pretty.Diff(curResult, result),
-			)
-		}
-
-		if pErr != nil {
-			// Initialize the error index.
-			pErr.SetErrorIndex(int32(index))
-
-			switch tErr := pErr.GetDetail().(type) {
-			case *roachpb.WriteTooOldError:
-				// We got a WriteTooOldError. We continue on to run all
-				// commands in the batch in order to determine the highest
-				// timestamp for more efficient retries. If the batch is
-				// transactional, we continue to lay down intents so that
-				// other concurrent overlapping transactions are forced
-				// through intent resolution and the chances of this batch
-				// succeeding when it will be retried are increased.
-				if writeTooOldErr != nil {
-					writeTooOldErr.GetDetail().(*roachpb.WriteTooOldError).ActualTimestamp.Forward(tErr.ActualTimestamp)
-				} else {
-					writeTooOldErr = pErr
-					// For transactions, we want to swallow the write too old error
-					// and just move the transaction timestamp forward and set the
-					// WriteTooOld flag. See below for exceptions.
-					if ba.Txn != nil {
-						returnWriteTooOldErr = false
-					}
-				}
-				// Set the flag to return a WriteTooOldError with the max timestamp
-				// encountered evaluating the entire batch on cput and inc requests.
-				// Because both of these requests must have their keys refreshed on
-				// commit with Transaction.WriteTooOld is true, and that refresh will
-				// fail, we'd be otherwise guaranteed to do a client-side retry.
-				// Returning an error allows a txn-coord-side retry.
-				switch args.(type) {
-				case *roachpb.ConditionalPutRequest:
-					// Conditional puts are an exception. Here, it makes less sense to
-					// continue because it's likely that the cput will fail on retry (a
-					// newer value is less likely to match the expected value). It's
-					// better to return the WriteTooOldError directly, allowing the txn
-					// coord sender to retry if it can refresh all other spans encountered
-					// already during the transaction, and then, if the cput results in a
-					// condition failed error, report that back to the client instead of a
-					// retryable error.
-					returnWriteTooOldErr = true
-				case *roachpb.IncrementRequest:
-					// Increments are an exception for similar reasons. If we wait until
-					// commit, we'll need a client-side retry, so we return immediately
-					// to see if we can do a txn coord sender retry instead.
-					returnWriteTooOldErr = true
-				case *roachpb.InitPutRequest:
-					// Init puts are also an exception. There's no reason to believe they
-					// will succeed on a retry, so better to short circuit and return the
-					// write too old error.
-					returnWriteTooOldErr = true
-				}
-				if ba.Txn != nil {
-					ba.Txn.Timestamp.Forward(tErr.ActualTimestamp)
-					ba.Txn.WriteTooOld = true
-				}
-				// Clear pErr; we're done processing it by having moved the
-				// batch or txn timestamps forward and set WriteTooOld if this
-				// is a transactional write. The EndTransaction will detect
-				// this pushed timestamp and return a TransactionRetryError.
-				pErr = nil
-			default:
-				return nil, result, pErr
-			}
-		}
-
-		if maxKeys != math.MaxInt64 {
-			retResults := reply.Header().NumKeys
-			if retResults > maxKeys {
-				log.Fatalf(ctx, "received %d results, limit was %d", retResults, maxKeys)
-			}
-			maxKeys -= retResults
-		}
-
-		// If transactional, we use ba.Txn for each individual command and
-		// accumulate updates to it.
-		// TODO(spencer,tschottdorf): need copy-on-write behavior for the
-		//   updated batch transaction / timestamp.
-		if ba.Txn != nil {
-			if txn := reply.Header().Txn; txn != nil {
-				ba.Txn.Update(txn)
-			}
-		}
-	}
-
-	// If there's a write too old error, return now that we've found
-	// the high water timestamp for retries.
-	if writeTooOldErr != nil && returnWriteTooOldErr {
-		return nil, result, writeTooOldErr
-	}
-
-	if ba.Txn != nil {
-		// If transactional, send out the final transaction entry with the reply.
-		br.Txn = ba.Txn
-		// If the transaction committed, forward the response
-		// timestamp to the commit timestamp in case we were able to
-		// optimize and commit at a higher timestamp without higher-level
-		// retry (i.e. there were no refresh spans and the commit timestamp
-		// wasn't leaked).
-		if ba.Txn.Status == roachpb.COMMITTED {
-			br.Timestamp.Forward(ba.Txn.Timestamp)
-		}
-	}
-	// Always update the batch response timestamp field to the timestamp at
-	// which the batch executed.
-	br.Timestamp.Forward(ba.Timestamp)
-
-	return br, result, nil
 }
 
 func (r *Replica) startKey() roachpb.RKey {
