+++ pkg/storage/replica_evaluate.go
@@ -1,0 +1,339 @@
+// Copyright 2019 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package storage
+
+import (
+	"bytes"
+	"context"
+	"math"
+
+	"github.com/cockroachdb/cockroach/pkg/roachpb"
+	"github.com/cockroachdb/cockroach/pkg/storage/batcheval"
+	"github.com/cockroachdb/cockroach/pkg/storage/batcheval/result"
+	"github.com/cockroachdb/cockroach/pkg/storage/engine"
+	"github.com/cockroachdb/cockroach/pkg/storage/engine/enginepb"
+	"github.com/cockroachdb/cockroach/pkg/storage/storagebase"
+	"github.com/cockroachdb/cockroach/pkg/util/log"
+	"github.com/kr/pretty"
+)
+
+// optimizePuts searches for contiguous runs of Put & CPut commands in
+// the supplied request union. Any run which exceeds a minimum length
+// threshold employs a full order iterator to determine whether the
+// range of keys being written is empty. If so, then the run can be
+// set to put "blindly", meaning no iterator need be used to read
+// existing values during the MVCC write.
+// The caller should use the returned slice (which is either equal to
+// the input slice, or has been shallow-copied appropriately to avoid
+// mutating the original requests).
+func optimizePuts(
+	batch engine.ReadWriter, origReqs []roachpb.RequestUnion, distinctSpans bool,
+) []roachpb.RequestUnion {
+	var minKey, maxKey roachpb.Key
+	var unique map[string]struct{}
+	if !distinctSpans {
+		unique = make(map[string]struct{}, len(origReqs))
+	}
+	// Returns false on occurrence of a duplicate key.
+	maybeAddPut := func(key roachpb.Key) bool {
+		// Note that casting the byte slice key to a string does not allocate.
+		if unique != nil {
+			if _, ok := unique[string(key)]; ok {
+				return false
+			}
+			unique[string(key)] = struct{}{}
+		}
+		if minKey == nil || bytes.Compare(key, minKey) < 0 {
+			minKey = key
+		}
+		if maxKey == nil || bytes.Compare(key, maxKey) > 0 {
+			maxKey = key
+		}
+		return true
+	}
+
+	firstUnoptimizedIndex := len(origReqs)
+	for i, r := range origReqs {
+		switch t := r.GetInner().(type) {
+		case *roachpb.PutRequest:
+			if maybeAddPut(t.Key) {
+				continue
+			}
+		case *roachpb.ConditionalPutRequest:
+			if maybeAddPut(t.Key) {
+				continue
+			}
+		case *roachpb.InitPutRequest:
+			if maybeAddPut(t.Key) {
+				continue
+			}
+		}
+		firstUnoptimizedIndex = i
+		break
+	}
+
+	if firstUnoptimizedIndex < optimizePutThreshold { // don't bother if below this threshold
+		return origReqs
+	}
+	iter := batch.NewIterator(engine.IterOptions{
+		// We want to include maxKey in our scan. Since UpperBound is exclusive, we
+		// need to set it to the key after maxKey.
+		UpperBound: maxKey.Next(),
+	})
+	defer iter.Close()
+
+	// If there are enough puts in the run to justify calling seek,
+	// we can determine whether any part of the range being written
+	// is "virgin" and set the puts to write blindly.
+	// Find the first non-empty key in the run.
+	iter.Seek(engine.MakeMVCCMetadataKey(minKey))
+	var iterKey roachpb.Key
+	if ok, err := iter.Valid(); err != nil {
+		// TODO(bdarnell): return an error here instead of silently
+		// running without the optimization?
+		log.Errorf(context.TODO(), "Seek returned error; disabling blind-put optimization: %s", err)
+		return origReqs
+	} else if ok && bytes.Compare(iter.Key().Key, maxKey) <= 0 {
+		iterKey = iter.Key().Key
+	}
+	// Set the prefix of the run which is being written to virgin
+	// keyspace to "blindly" put values.
+	reqs := append([]roachpb.RequestUnion(nil), origReqs...)
+	for i := range reqs[:firstUnoptimizedIndex] {
+		inner := reqs[i].GetInner()
+		if iterKey == nil || bytes.Compare(iterKey, inner.Header().Key) > 0 {
+			switch t := inner.(type) {
+			case *roachpb.PutRequest:
+				shallow := *t
+				shallow.Blind = true
+				reqs[i].MustSetInner(&shallow)
+			case *roachpb.ConditionalPutRequest:
+				shallow := *t
+				shallow.Blind = true
+				reqs[i].MustSetInner(&shallow)
+			case *roachpb.InitPutRequest:
+				shallow := *t
+				shallow.Blind = true
+				reqs[i].MustSetInner(&shallow)
+			default:
+				log.Fatalf(context.TODO(), "unexpected non-put request: %s", t)
+			}
+		}
+	}
+	return reqs
+}
+
+// evaluateBatch evaluates a batch request by splitting it up into its
+// individual commands, passing them to evaluateCommand, and combining
+// the results.
+func evaluateBatch(
+	ctx context.Context,
+	idKey storagebase.CmdIDKey,
+	batch engine.ReadWriter,
+	rec batcheval.EvalContext,
+	ms *enginepb.MVCCStats,
+	ba roachpb.BatchRequest,
+) (*roachpb.BatchResponse, result.Result, *roachpb.Error) {
+	br := ba.CreateReply()
+
+	maxKeys := int64(math.MaxInt64)
+	if ba.Header.MaxSpanRequestKeys != 0 {
+		// We have a batch of requests with a limit. We keep track of how many
+		// remaining keys we can touch.
+		maxKeys = ba.Header.MaxSpanRequestKeys
+	}
+
+	// Optimize any contiguous sequences of put and conditional put ops.
+	if len(ba.Requests) >= optimizePutThreshold {
+		ba.Requests = optimizePuts(batch, ba.Requests, ba.Header.DistinctSpans)
+	}
+
+	// Create a shallow clone of the transaction. We only modify a few
+	// non-pointer fields (Sequence, DeprecatedBatchIndex, WriteTooOld, Timestamp),
+	// so this saves a few allocs.
+	if ba.Txn != nil {
+		txnShallow := *ba.Txn
+		ba.Txn = &txnShallow
+
+		// Check whether this transaction has been aborted, if applicable.
+		// This applies to writes that leave intents (the use of the
+		// IsTransactionWrite flag excludes operations like HeartbeatTxn),
+		// and reads that occur in a transaction that has already written
+		// (see #2231 for more about why we check for aborted transactions
+		// on reads). Note that 1PC transactions have had their
+		// transaction field cleared by this point so we do not execute
+		// this check in that case.
+		if ba.IsTransactionWrite() || ba.Txn.Writing {
+			// We don't check the abort span for a couple of special requests:
+			// - if the request is asking to abort the transaction, then don't check the
+			// AbortSpan; we don't want the request to be rejected if the transaction
+			// has already been aborted.
+			// - heartbeats don't check the abort span. If the txn is aborted, they'll
+			// return an aborted proto in their otherwise successful response.
+			singleAbort := ba.IsSingleEndTransactionRequest() &&
+				!ba.Requests[0].GetInner().(*roachpb.EndTransactionRequest).Commit
+			if !singleAbort && !ba.IsSingleHeartbeatTxnRequest() {
+				if pErr := checkIfTxnAborted(ctx, rec, batch, *ba.Txn); pErr != nil {
+					return nil, result.Result{}, pErr
+				}
+			}
+		}
+	}
+
+	var result result.Result
+	var writeTooOldErr *roachpb.Error
+	returnWriteTooOldErr := true
+
+	for index, union := range ba.Requests {
+		// Execute the command.
+		args := union.GetInner()
+		if ba.Txn != nil {
+			// Sequence numbers used to be set on each BatchRequest instead of
+			// on each individual Request. This meant that all Requests in a
+			// BatchRequest shared the same sequence number, so a BatchIndex was
+			// augmented to provide an ordering between them. Individual
+			// Requests were later given their own sequence numbers, so the
+			// BatchIndex was no longer necessary.
+			if seqNum := args.Header().Sequence; seqNum != 0 {
+				// Set the Request's sequence number on the TxnMeta for this
+				// request. Each request will set their own sequence number on
+				// the TxnMeta, which is stored as part of an intent.
+				ba.Txn.Sequence = seqNum
+			}
+		}
+		// Note that responses are populated even when an error is returned.
+		// TODO(tschottdorf): Change that. IIRC there is nontrivial use of it currently.
+		reply := br.Responses[index].GetInner()
+		curResult, pErr := evaluateCommand(ctx, idKey, index, batch, rec, ms, ba.Header, maxKeys, args, reply)
+
+		if err := result.MergeAndDestroy(curResult); err != nil {
+			// TODO(tschottdorf): see whether we really need to pass nontrivial
+			// Result up on error and if so, formalize that.
+			log.Fatalf(
+				ctx,
+				"unable to absorb Result: %s\ndiff(new, old): %s",
+				err, pretty.Diff(curResult, result),
+			)
+		}
+
+		if pErr != nil {
+			// Initialize the error index.
+			pErr.SetErrorIndex(int32(index))
+
+			switch tErr := pErr.GetDetail().(type) {
+			case *roachpb.WriteTooOldError:
+				// We got a WriteTooOldError. We continue on to run all
+				// commands in the batch in order to determine the highest
+				// timestamp for more efficient retries. If the batch is
+				// transactional, we continue to lay down intents so that
+				// other concurrent overlapping transactions are forced
+				// through intent resolution and the chances of this batch
+				// succeeding when it will be retried are increased.
+				if writeTooOldErr != nil {
+					writeTooOldErr.GetDetail().(*roachpb.WriteTooOldError).ActualTimestamp.Forward(tErr.ActualTimestamp)
+				} else {
+					writeTooOldErr = pErr
+					// For transactions, we want to swallow the write too old error
+					// and just move the transaction timestamp forward and set the
+					// WriteTooOld flag. See below for exceptions.
+					if ba.Txn != nil {
+						returnWriteTooOldErr = false
+					}
+				}
+				// Set the flag to return a WriteTooOldError with the max timestamp
+				// encountered evaluating the entire batch on cput and inc requests.
+				// Because both of these requests must have their keys refreshed on
+				// commit with Transaction.WriteTooOld is true, and that refresh will
+				// fail, we'd be otherwise guaranteed to do a client-side retry.
+				// Returning an error allows a txn-coord-side retry.
+				switch args.(type) {
+				case *roachpb.ConditionalPutRequest:
+					// Conditional puts are an exception. Here, it makes less sense to
+					// continue because it's likely that the cput will fail on retry (a
+					// newer value is less likely to match the expected value). It's
+					// better to return the WriteTooOldError directly, allowing the txn
+					// coord sender to retry if it can refresh all other spans encountered
+					// already during the transaction, and then, if the cput results in a
+					// condition failed error, report that back to the client instead of a
+					// retryable error.
+					returnWriteTooOldErr = true
+				case *roachpb.IncrementRequest:
+					// Increments are an exception for similar reasons. If we wait until
+					// commit, we'll need a client-side retry, so we return immediately
+					// to see if we can do a txn coord sender retry instead.
+					returnWriteTooOldErr = true
+				case *roachpb.InitPutRequest:
+					// Init puts are also an exception. There's no reason to believe they
+					// will succeed on a retry, so better to short circuit and return the
+					// write too old error.
+					returnWriteTooOldErr = true
+				}
+				if ba.Txn != nil {
+					ba.Txn.Timestamp.Forward(tErr.ActualTimestamp)
+					ba.Txn.WriteTooOld = true
+				}
+				// Clear pErr; we're done processing it by having moved the
+				// batch or txn timestamps forward and set WriteTooOld if this
+				// is a transactional write. The EndTransaction will detect
+				// this pushed timestamp and return a TransactionRetryError.
+				pErr = nil
+			default:
+				return nil, result, pErr
+			}
+		}
+
+		if maxKeys != math.MaxInt64 {
+			retResults := reply.Header().NumKeys
+			if retResults > maxKeys {
+				log.Fatalf(ctx, "received %d results, limit was %d", retResults, maxKeys)
+			}
+			maxKeys -= retResults
+		}
+
+		// If transactional, we use ba.Txn for each individual command and
+		// accumulate updates to it.
+		// TODO(spencer,tschottdorf): need copy-on-write behavior for the
+		//   updated batch transaction / timestamp.
+		if ba.Txn != nil {
+			if txn := reply.Header().Txn; txn != nil {
+				ba.Txn.Update(txn)
+			}
+		}
+	}
+
+	// If there's a write too old error, return now that we've found
+	// the high water timestamp for retries.
+	if writeTooOldErr != nil && returnWriteTooOldErr {
+		return nil, result, writeTooOldErr
+	}
+
+	if ba.Txn != nil {
+		// If transactional, send out the final transaction entry with the reply.
+		br.Txn = ba.Txn
+		// If the transaction committed, forward the response
+		// timestamp to the commit timestamp in case we were able to
+		// optimize and commit at a higher timestamp without higher-level
+		// retry (i.e. there were no refresh spans and the commit timestamp
+		// wasn't leaked).
+		if ba.Txn.Status == roachpb.COMMITTED {
+			br.Timestamp.Forward(ba.Txn.Timestamp)
+		}
+	}
+	// Always update the batch response timestamp field to the timestamp at
+	// which the batch executed.
+	br.Timestamp.Forward(ba.Timestamp)
+
+	return br, result, nil
+}
