+++ pkg/storage/replica_read.go
@@ -1,0 +1,155 @@
+// Copyright 2019 The Cockroach Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+// implied. See the License for the specific language governing
+// permissions and limitations under the License.
+
+package storage
+
+import (
+	"context"
+
+	"github.com/cockroachdb/cockroach/pkg/keys"
+	"github.com/cockroachdb/cockroach/pkg/roachpb"
+	"github.com/cockroachdb/cockroach/pkg/storage/batcheval/result"
+	"github.com/cockroachdb/cockroach/pkg/storage/closedts/ctpb"
+	"github.com/cockroachdb/cockroach/pkg/storage/closedts/storage"
+	"github.com/cockroachdb/cockroach/pkg/storage/spanset"
+	"github.com/cockroachdb/cockroach/pkg/storage/storagebase"
+	"github.com/cockroachdb/cockroach/pkg/storage/storagepb"
+	"github.com/cockroachdb/cockroach/pkg/util"
+	"github.com/cockroachdb/cockroach/pkg/util/log"
+)
+
+// executeReadOnlyBatch updates the read timestamp cache and waits for any
+// overlapping writes currently processing through Raft ahead of us to
+// clear via the latches.
+func (r *Replica) executeReadOnlyBatch(
+	ctx context.Context, ba roachpb.BatchRequest,
+) (br *roachpb.BatchResponse, pErr *roachpb.Error) {
+	// If the read is not inconsistent, the read requires the range lease or
+	// permission to serve via follower reads.
+	var status storagepb.LeaseStatus
+	if ba.ReadConsistency.RequiresReadLease() {
+		if status, pErr = r.redirectOnOrAcquireLease(ctx); pErr != nil {
+			if lErr, ok := pErr.GetDetail().(*roachpb.NotLeaseHolderError); ok &&
+				FollowerReadsEnabled.Get(&r.store.cfg.Settings.SV) &&
+				lErr.LeaseHolder != nil && lErr.Lease.Type() == roachpb.LeaseEpoch {
+
+				r.mu.RLock()
+				lai := r.mu.state.LeaseAppliedIndex
+				r.mu.RUnlock()
+				if !r.store.cfg.ClosedTimestamp.Provider.CanServe(
+					lErr.LeaseHolder.NodeID, ba.Timestamp, r.RangeID, ctpb.Epoch(lErr.Lease.Epoch), ctpb.LAI(lai),
+				) {
+					r.store.cfg.ClosedTimestamp.Clients.Request(lErr.LeaseHolder.NodeID, r.RangeID)
+					// NB: this can't go behind V(x) because the log message created by the
+					// storage might be gigantic in real clusters, and we don't want to trip it
+					// using logspy.
+					if false {
+						log.Warningf(ctx, "can't serve follower read for %s at epo %d lai %d, storage is %s",
+							ba.Timestamp, lErr.Lease.Epoch, lai,
+							r.store.cfg.ClosedTimestamp.Storage.(*storage.MultiStorage).StringForNodes(lErr.LeaseHolder.NodeID),
+						)
+					}
+					return nil, pErr
+				}
+				// This replica can serve this read!
+				//
+				// TODO(tschottdorf): once a read for a timestamp T has been served, the replica may
+				// serve reads for that and smaller timestamps forever.
+				log.Event(ctx, "serving via follower read")
+			} else {
+				return nil, pErr
+			}
+		}
+	}
+	r.limitTxnMaxTimestamp(ctx, &ba, status)
+
+	spans, err := r.collectSpans(&ba)
+	if err != nil {
+		return nil, roachpb.NewError(err)
+	}
+
+	// Acquire latches to prevent overlapping commands from executing
+	// until this command completes.
+	log.Event(ctx, "acquire latches")
+	endCmds, err := r.beginCmds(ctx, &ba, spans)
+	if err != nil {
+		return nil, roachpb.NewError(err)
+	}
+
+	log.Event(ctx, "waiting for read lock")
+	r.readOnlyCmdMu.RLock()
+	defer r.readOnlyCmdMu.RUnlock()
+
+	// Guarantee we release the latches that we just acquired. It is
+	// important that this is inside the readOnlyCmdMu lock so that the
+	// timestamp cache update is synchronized. This is wrapped to delay
+	// pErr evaluation to its value when returning.
+	defer func() {
+		endCmds.done(br, pErr, proposalNoReevaluation)
+	}()
+
+	// TODO(nvanbenschoten): Can this be moved into Replica.requestCanProceed?
+	if _, err := r.IsDestroyed(); err != nil {
+		return nil, roachpb.NewError(err)
+	}
+
+	rSpan, err := keys.Range(ba)
+	if err != nil {
+		return nil, roachpb.NewError(err)
+	}
+
+	if err := r.requestCanProceed(rSpan, ba.Timestamp); err != nil {
+		return nil, roachpb.NewError(err)
+	}
+
+	// Evaluate read-only batch command. It checks for matching key range; note
+	// that holding readOnlyCmdMu throughout is important to avoid reads from the
+	// "wrong" key range being served after the range has been split.
+	var result result.Result
+	rec := NewReplicaEvalContext(r, spans)
+	readOnly := r.store.Engine().NewReadOnly()
+	if util.RaceEnabled {
+		readOnly = spanset.NewReadWriter(readOnly, spans)
+	}
+	defer readOnly.Close()
+	br, result, pErr = evaluateBatch(ctx, storagebase.CmdIDKey(""), readOnly, rec, nil, ba)
+
+	if result.Local.DetachMaybeWatchForMerge() {
+		if err := r.maybeWatchForMerge(ctx); err != nil {
+			return nil, roachpb.NewError(err)
+		}
+	}
+
+	if intents := result.Local.DetachIntents(); len(intents) > 0 {
+		log.Eventf(ctx, "submitting %d intents to asynchronous processing", len(intents))
+		// We only allow synchronous intent resolution for consistent requests.
+		// Intent resolution is async/best-effort for inconsistent requests.
+		//
+		// An important case where this logic is necessary is for RangeLookup
+		// requests. In their case, synchronous intent resolution can deadlock
+		// if the request originated from the local node which means the local
+		// range descriptor cache has an in-flight RangeLookup request which
+		// prohibits any concurrent requests for the same range. See #17760.
+		allowSyncProcessing := ba.ReadConsistency == roachpb.CONSISTENT
+		if err := r.store.intentResolver.cleanupIntentsAsync(ctx, r, intents, allowSyncProcessing); err != nil {
+			log.Warning(ctx, err)
+		}
+	}
+	if pErr != nil {
+		log.VErrEvent(ctx, 3, pErr.String())
+	} else {
+		log.Event(ctx, "read completed")
+	}
+	return br, pErr
+}
